+++ CHANGELOG.md
@@ -5,3 +5,10 @@ Version 1.0.0 *(2018-03-27)*
 ------------------------------
 
 * Initial release.
+
+Version 1.0.1 *(2018-05-18)*
+------------------------------
+
+* Downgraded minimum SDK version from 17 to 15
+* Deprecated `areFieldsValidRecursive(ViewGroup)` method in favor of `validateLayout(View)`
+* Updated dependencies to latest version